+++ core/src/wallet/ethereum/EthereumLikeAccount.cpp
@@ -46,6 +46,7 @@
 #include <wallet/ethereum/ERC20/ERC20LikeOperation.h>
 #include <wallet/common/database/BlockDatabaseHelper.h>
 #include <wallet/pool/database/CurrenciesDatabaseHelper.hpp>
+#include <wallet/pool/WalletPool.hpp>
 #include <events/Event.hpp>
 #include <math/Base58.hpp>
 #include <utils/Option.hpp>
@@ -333,19 +334,26 @@ namespace ledger {
         }
 
         FuturePtr<Amount> EthereumLikeAccount::getBalance() {
+            auto cachedBalance = getWallet()->getBalanceFromCache(getIndex());
+            if (cachedBalance.hasValue()) {
+                return FuturePtr<Amount>::successful(std::make_shared<Amount>(cachedBalance.getValue()));
+            }
             std::vector<EthereumLikeKeychain::Address> listAddresses{_keychain->getAddress()};
-                auto currency = getWallet()->getCurrency();
-                return _explorer->getBalance(listAddresses).mapPtr<Amount>(getContext(), [currency] (const std::shared_ptr<BigInt> &balance) -> std::shared_ptr<Amount> {
-                    return std::make_shared<Amount>(currency, 0, BigInt(balance->toString()));
-                });
+            auto currency = getWallet()->getCurrency();
+            auto self = getSelf();
+            return _explorer->getBalance(listAddresses).mapPtr<Amount>(getMainExecutionContext(), [self, currency] (const std::shared_ptr<BigInt> &balance) -> std::shared_ptr<Amount> {
+                Amount b(currency, 0, BigInt(balance->toString()));
+                self->getWallet()->updateBalanceCache(self->getIndex(), b);
+                return std::make_shared<Amount>(b);
+            });
         }
 
         std::shared_ptr<api::OperationQuery> EthereumLikeAccount::queryOperations() {
             auto query = std::make_shared<OperationQuery>(
                     api::QueryFilter::accountEq(getAccountUid()),
                     getWallet()->getDatabase(),
-                    getWallet()->getContext(),
-                    getWallet()->getMainExecutionContext()
+                    getWallet()->getPool()->getThreadPoolExecutionContext(),
+                    getMainExecutionContext()
             );
             query->registerAccount(shared_from_this());
             return query;
@@ -364,7 +372,7 @@ namespace ledger {
                                                const std::string & end,
                                                api::TimePeriod precision) {
                 auto self = getSelf();
-                return async<std::vector<std::shared_ptr<api::Amount>>>([=] () -> std::vector<std::shared_ptr<api::Amount>> {
+                return Future<std::vector<std::shared_ptr<api::Amount>>>::async(getWallet()->getPool()->getThreadPoolExecutionContext(), [=] () -> std::vector<std::shared_ptr<api::Amount>> {
                     auto startDate = DateUtils::fromJSON(start);
                     auto endDate = DateUtils::fromJSON(end);
                     if (startDate >= endDate) {
@@ -480,43 +488,42 @@ namespace ledger {
         }
 
         std::shared_ptr<api::EventBus> EthereumLikeAccount::synchronize() {
-                std::lock_guard<std::mutex> lock(_synchronizationLock);
-                if (_currentSyncEventBus)
-                        return _currentSyncEventBus;
-                auto eventPublisher = std::make_shared<EventPublisher>(getContext());
+            std::lock_guard<std::mutex> lock(_synchronizationLock);
+            if (_currentSyncEventBus)
+                    return _currentSyncEventBus;
+            auto eventPublisher = std::make_shared<EventPublisher>(getContext());
 
-                _currentSyncEventBus = eventPublisher->getEventBus();
-                auto future = _synchronizer->synchronize(std::static_pointer_cast<EthereumLikeAccount>(shared_from_this()))->getFuture();
-                auto self = std::static_pointer_cast<EthereumLikeAccount>(shared_from_this());
+            _currentSyncEventBus = eventPublisher->getEventBus();
+            auto future = _synchronizer->synchronize(std::static_pointer_cast<EthereumLikeAccount>(shared_from_this()))->getFuture();
+            auto self = std::static_pointer_cast<EthereumLikeAccount>(shared_from_this());
 
-                //Update current block height (needed to compute trust level)
-                _explorer->getCurrentBlock().onComplete(getContext(), [self] (const TryPtr<EthereumLikeBlockchainExplorer::Block>& block) mutable {
-                    if (block.isSuccess()) {
-                            self->_currentBlockHeight = block.getValue()->height;
-                    }
-                });
+            //Update current block height (needed to compute trust level)
+            _explorer->getCurrentBlock().onComplete(getContext(), [self] (const TryPtr<EthereumLikeBlockchainExplorer::Block>& block) mutable {
+                if (block.isSuccess()) {
+                    self->_currentBlockHeight = block.getValue()->height;
+                }
+            });
 
-                auto startTime = DateUtils::now();
-                eventPublisher->postSticky(std::make_shared<Event>(api::EventCode::SYNCHRONIZATION_STARTED, api::DynamicObject::newInstance()), 0);
-                future.onComplete(getContext(), [eventPublisher, self, startTime] (const Try<Unit>& result) {
-                    api::EventCode code;
-                    auto payload = std::make_shared<DynamicObject>();
-                    auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(DateUtils::now() - startTime).count();
-                    payload->putLong(api::Account::EV_SYNC_DURATION_MS, duration);
-                    if (result.isSuccess()) {
-                            code = api::EventCode::SYNCHRONIZATION_SUCCEED;
-                    } else {
-                            code = api::EventCode::SYNCHRONIZATION_FAILED;
-                            payload->putString(api::Account::EV_SYNC_ERROR_CODE, api::to_string(result.getFailure().getErrorCode()));
-                            payload->putInt(api::Account::EV_SYNC_ERROR_CODE_INT, (int32_t)result.getFailure().getErrorCode());
-                            payload->putString(api::Account::EV_SYNC_ERROR_MESSAGE, result.getFailure().getMessage());
-                    }
-                    eventPublisher->postSticky(std::make_shared<Event>(code, payload), 0);
-                    std::lock_guard<std::mutex> lock(self->_synchronizationLock);
-                    self->_currentSyncEventBus = nullptr;
-
-                });
-                return eventPublisher->getEventBus();
+            auto startTime = DateUtils::now();
+            eventPublisher->postSticky(std::make_shared<Event>(api::EventCode::SYNCHRONIZATION_STARTED, api::DynamicObject::newInstance()), 0);
+            future.onComplete(getContext(), [eventPublisher, self, startTime] (const Try<Unit>& result) {
+                api::EventCode code;
+                auto payload = std::make_shared<DynamicObject>();
+                auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(DateUtils::now() - startTime).count();
+                payload->putLong(api::Account::EV_SYNC_DURATION_MS, duration);
+                if (result.isSuccess()) {
+                    code = api::EventCode::SYNCHRONIZATION_SUCCEED;
+                } else {
+                    code = api::EventCode::SYNCHRONIZATION_FAILED;
+                    payload->putString(api::Account::EV_SYNC_ERROR_CODE, api::to_string(result.getFailure().getErrorCode()));
+                    payload->putInt(api::Account::EV_SYNC_ERROR_CODE_INT, (int32_t)result.getFailure().getErrorCode());
+                    payload->putString(api::Account::EV_SYNC_ERROR_MESSAGE, result.getFailure().getMessage());
+                }
+                eventPublisher->postSticky(std::make_shared<Event>(code, payload), 0);
+                std::lock_guard<std::mutex> lock(self->_synchronizationLock);
+                self->_currentSyncEventBus = nullptr;
+            });
+            return eventPublisher->getEventBus();
         }
 
         std::shared_ptr<EthereumLikeAccount> EthereumLikeAccount::getSelf() {
@@ -595,7 +602,7 @@ namespace ledger {
                 });
 
                 return txHash;
-            }).callback(getContext(), callback);
+            }).callback(getMainExecutionContext(), callback);
         }
 
         void EthereumLikeAccount::broadcastTransaction(const std::shared_ptr<api::EthereumLikeTransaction> & transaction,
@@ -613,26 +620,41 @@ namespace ledger {
         }
 
         void EthereumLikeAccount::getGasPrice(const std::shared_ptr<api::BigIntCallback> & callback) {
-            _explorer->getGasPrice().mapPtr<api::BigInt>(getContext(), [] (const std::shared_ptr<BigInt> &gasPrice) -> std::shared_ptr<api::BigInt> {
+            _explorer->getGasPrice().mapPtr<api::BigInt>(getMainExecutionContext(), [] (const std::shared_ptr<BigInt> &gasPrice) -> std::shared_ptr<api::BigInt> {
                 return std::make_shared<api::BigIntImpl>(*gasPrice);
-            }).callback(getContext(), callback);
+            }).callback(getMainExecutionContext(), callback);
         }
 
         void EthereumLikeAccount::getEstimatedGasLimit(const std::string & address, const std::shared_ptr<api::BigIntCallback> & callback) {
-            _explorer->getEstimatedGasLimit(address).mapPtr<api::BigInt>(getContext(), [] (const std::shared_ptr<BigInt> &gasPrice) -> std::shared_ptr<api::BigInt> {
+            _explorer->getEstimatedGasLimit(address).mapPtr<api::BigInt>(getMainExecutionContext(), [] (const std::shared_ptr<BigInt> &gasPrice) -> std::shared_ptr<api::BigInt> {
                 return std::make_shared<api::BigIntImpl>(*gasPrice);
-            }).callback(getContext(), callback);
+            }).callback(getMainExecutionContext(), callback);
         }
 
         FuturePtr<api::BigInt> EthereumLikeAccount::getERC20Balance(const std::string & erc20Address) {
-            return _explorer->getERC20Balance(_keychain->getAddress()->toEIP55(), erc20Address).mapPtr<api::BigInt>(getContext(), [] (const std::shared_ptr<BigInt> &erc20Balance) -> std::shared_ptr<api::BigInt> {
+            return _explorer->getERC20Balance(_keychain->getAddress()->toEIP55(), erc20Address).mapPtr<api::BigInt>(getMainExecutionContext(), [] (const std::shared_ptr<BigInt> &erc20Balance) -> std::shared_ptr<api::BigInt> {
                 return std::make_shared<api::BigIntImpl>(*erc20Balance);
             });
         }
 
         void EthereumLikeAccount::getERC20Balance(const std::string & erc20Address,
                                                   const std::shared_ptr<api::BigIntCallback> & callback) {
-            getERC20Balance(erc20Address).callback(getContext(), callback);
+            getERC20Balance(erc20Address).callback(getMainExecutionContext(), callback);
+        }
+
+        Future<std::vector<std::shared_ptr<api::BigInt>>> EthereumLikeAccount::getERC20Balances(const std::vector<std::string> &erc20Addresses) {
+            return _explorer->getERC20Balances(_keychain->getAddress()->toEIP55(), erc20Addresses)
+                    .map<std::vector<std::shared_ptr<api::BigInt>>>(getMainExecutionContext(),
+                                                                    [] (const std::vector<BigInt> &erc20Balances) {
+                                                                        return vector::map<std::shared_ptr<api::BigInt>, BigInt>(erc20Balances, [] (const BigInt &erc20Balance) {
+                                                                            return std::make_shared<api::BigIntImpl>(erc20Balance);
+                                                                        });
+                                                                    }
+                    );
+        }
+
+        void EthereumLikeAccount::getERC20Balances(const std::vector<std::string> &erc20Addresses, const std::shared_ptr<api::BigIntListCallback> & callback) {
+            getERC20Balances(erc20Addresses).callback(getMainExecutionContext(), callback);
         }
 
         void EthereumLikeAccount::addERC20Accounts(soci::session &sql,
@@ -653,7 +675,7 @@ namespace ledger {
                 auto self = std::dynamic_pointer_cast<EthereumLikeAccount>(shared_from_this());
                 auto buildFunction = [self] (const EthereumLikeTransactionBuildRequest& request, const std::shared_ptr<EthereumLikeBlockchainExplorer> &explorer) -> Future<std::shared_ptr<api::EthereumLikeTransaction>> {
                     // Check if balance is sufficient
-                    return self->getBalance().flatMapPtr<api::EthereumLikeTransaction>(self->getContext(), [self, request, explorer](const std::shared_ptr<Amount> &balance) {
+                    return self->getBalance().flatMapPtr<api::EthereumLikeTransaction>(self->getMainExecutionContext(), [self, request, explorer](const std::shared_ptr<Amount> &balance) {
                         // Check if all needed values are set
                         if (!request.gasLimit || !request.gasPrice || (!request.value && !request.wipe)) {
                             throw make_exception(api::ErrorCode::INVALID_ARGUMENT, "Missing mandatory informations (e.g. gasLimit, gasPrice or value).");
@@ -672,14 +694,14 @@ namespace ledger {
                         tx->setReceiver(request.toAddress);
                         auto accountAddress = self->getKeychain()->getAddress()->toString();
                         tx->setSender(accountAddress);
-                        return explorer->getNonce(accountAddress).map<std::shared_ptr<api::EthereumLikeTransaction>>(self->getContext(), [self, tx] (const std::shared_ptr<BigInt> &nonce) -> std::shared_ptr<api::EthereumLikeTransaction> {
+                        return explorer->getNonce(accountAddress).map<std::shared_ptr<api::EthereumLikeTransaction>>(self->getMainExecutionContext(), [self, tx] (const std::shared_ptr<BigInt> &nonce) -> std::shared_ptr<api::EthereumLikeTransaction> {
                             tx->setNonce(nonce);
                             return tx;
                         });
                     });
                 };
 
-                return std::make_shared<EthereumLikeTransactionBuilder>(getContext(),
+                return std::make_shared<EthereumLikeTransactionBuilder>(getMainExecutionContext(),
                                                                         getWallet()->getCurrency(),
                                                                         _explorer,
                                                                         logger(),