+++ core/src/wallet/ethereum/api_impl/EthereumLikeOperation.h
@@ -47,6 +47,9 @@ namespace ledger {
         private:
             std::shared_ptr<api::EthereumLikeTransaction> _transaction;
             std::vector<std::shared_ptr<api::InternalTransaction>> _internalTxs;
+            std::shared_ptr<OperationApi> _backend;
+            // To allow lazy loading of internal transactions
+            bool _internalTxsRetrieved;
         };
 
     }