+++ core/src/wallet/ethereum/database/EthereumLikeTransactionDatabaseHelper.cpp
@@ -106,8 +106,8 @@ namespace ledger {
         }
 
         std::string EthereumLikeTransactionDatabaseHelper::putTransaction(soci::session &sql,
-                                                                         const std::string& accountUid,
-                                                                         const EthereumLikeBlockchainExplorerTransaction &tx) {
+                                                                          const std::string &accountUid,
+                                                                          const EthereumLikeBlockchainExplorerTransaction &tx) {
             auto blockUid = tx.block.map<std::string>([] (const EthereumLikeBlockchainExplorer::Block& block) {
                 return block.getUid();
             });