+++ core/src/wallet/ethereum/explorers/LedgerApiEthereumLikeBlockchainExplorer.h
@@ -55,7 +55,14 @@ namespace ledger {
             Future<std::shared_ptr<BigInt>> getBalance(const std::vector<EthereumLikeKeychain::Address> &addresses) override;
             Future<std::shared_ptr<BigInt>> getGasPrice() override;
             Future<std::shared_ptr<BigInt>> getEstimatedGasLimit(const std::string &address) override;
-            Future<std::shared_ptr<BigInt>> getERC20Balance(const std::string &address, const std::string &erc20Address) override;
+
+            Future<std::shared_ptr<BigInt>>
+            getERC20Balance(const std::string &address,
+                            const std::string &erc20Address) override;
+            Future<std::vector<BigInt>>
+            getERC20Balances(const std::string &address,
+                             const std::vector<std::string> &erc20Addresses) override;
+
             Future<String> pushLedgerApiTransaction(const std::vector<uint8_t> &transaction) override;
             Future<void *> startSession() override;
             Future<Unit> killSession(void *session) override;